Add ability to short-circuit the preorder traversal
Add an overload to the TreeWalkerExtensions.PreOrderTraversal method to allow short-circuiting of the traversal by excluding subtrees from the traversal.  The overload includes a function called 'excludeSubtreePredicate' that takes the current node being evaluated and the depth of the node from the original node that the traversal began on and returns a boolean.  If the function returns true then the node and all of its descendants will be excluded from the traversal.