Remove 'includeNode' parameter from TreeWalkerExtensions.PostOrderTraversal
Remove the 'includeNode' parameter from all overloads of PostOrderTraversal.  This parameter allowed the user to include or exclude the node that the traversal began from.  However, it really does not make any sense not to include the beginning node when performing a traversal so the node is now always included.