Update TreeWalkerExtensions.LevelOrderTraversal
Remove the 'includeNode' parameter from all overloads of LevelOrderTraversal.  This parameter allowed the user to include or exclude the node that the traversal began from.  However, it really does not make any sense not to include the beginning node when performing a traversal so the node is now always included.

Add an overload to the TreeWalkerExtensions.LevelOrderTraversal method to allow short-circuiting of the traversal by excluding subtrees from the traversal.  The overload includes a function called 'excludeSubtreePredicate' that takes the current node being evaluated and the depth of the node from the original node that the traversal began on and returns a boolean.  If the function returns true then the node and all of its descendants will be excluded from the traversal.