Modify the IEnumerable<VirtualTree<T>>.GetChildAt() extension
Modify the IEnumerable<VirtualTree<T>>.GetChildAt() extension so that it does not throw an argument out of range exception.
The ArgumentOutOfRange exception makes sense when trying to get the child at a particular index for a single VirtualTree because there is no other way to express that we failed to get the child. But for IEnumerable<VirtualTree<T>>.GetChildAt() if we fail to get a child we can simply not add an item to the resulting IEnumerable. If we get no children can just return an empty IEnumerable.

closes #5