Change VirtualTree traversal return values
Modify VirtualTree traversal results to return IEnumerable<T> rather than IEnumerable<VirtualTree<T>>. The latter is more expensive for very large trees. Additionally, it is unlikely the consumer would ever need to materialize every descendant in a tree as a VirtualTree. In the off chance that they do it is easy enough for them to use Select to do so.

closes #2