Renamed several methods and types
VirtualTree.Create to VirtualTree.New:
I like 'new' better.  It is consistent with the C# keyword.

VirtualTree.CreateFromSelf to VirtualTree.ShallowCopy
I think this name is a little more clear.

SelectChildren to GetChildren
In LINQ 'Select' and 'SelectMany' both transform or project an object.  Same with SQL.  'SelectChildren', on the other hand, does not transform anything; it is a filter, more like a combination of from and where, really.  I am changing it from select to get because I think the behavior of 'SelectChildren' is inconsistent with LINQ.

SelectDescendants to GetDescendants
See argument for SelectChildren.

ExcludeOptions.ExcludeChildren to ExcludeOptions.ExcludeDescendants
Exclude descendants is a more accurate description of this option.