Replaced ITreeWalker.TryGetParent with ITreeWalker.GetAncestors
I believe trees can be divided into two groups: those that that can yield their ancestors in O(n) runtime and those that can yield their ancestors in O(1) runtime.  Given an ITreeWalker that implements TryGetParent the tree walker must get the node's ancestors by iteratively calling TryGetParent, resulting in an implementation of GetAncestors that has a runtime of O(n).  This is the case even if the tree itself natively supports getting its ancestors in O(1) runtime.  But given an ITreeWalker that implements GetAncestors the ITreeWalker can get a node's ancestors in O(1) runtime if the tree natively supports that; if the tree does not support a O(1) runtime then the tree walker will fall back to a O(n) runtime which is no worse than an ITreeWalker that implements TryGetParent.  In all cases, the runtime to get a parent is O(1).

closes #3